pyproject.toml
[tool.poetry]
name = "startapp"
version = "0.1.0"
description = ""
authors = ["kznar <kznartee16reg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
psycopg2 = "^2.9.2"
djangorestframework = "^3.13.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
